services/processor-service/requirements.txt
pika
python-dotenv
python-json-logger
requests